pyproject.toml
[tool.poetry]
name = "multiplayer-ai"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.12.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
